feat: move websocket instance to separate file
